+++ price_feeder/logger.py
@@ -1,0 +1,9 @@
+import logging
+import logging.config
+
+
+logging.basicConfig(level=logging.INFO,
+                    format='%(asctime)s %(name)-12s %(levelname)-8s %(message)s',
+                    datefmt='%Y-%m-%d %H:%M:%S')
+
+log = logging.getLogger('default')